Support plocate alternative to mlocate
Declare Fedora 36 support.

The parameter `mlocate_packages` is deprecated.

Setting it to `plocate` or anything other than
the default `mlocate` never worked anyway.

Install and use `plocate` always on Fedora 37 and 38.

While `plocate` has been available in EL8 or Fedora 36 in
Fedora 37 and newer the `plocate` RPM obsoletes
the `mlocate` package.

The default for EL8, EL9 and Fedora 36 remains `mlocate` however
the new parameter `mlocate` can be set to `plocate` to install
and configure `plocate`

* https://fedoraproject.org/wiki/Changes/Plocate_as_the_default_locate_implementation